+++ integration_tests/test_upgrade.py
@@ -184,6 +184,9 @@ def test_cosmovisor_upgrade(custom_cronos: Cronos, tmp_path_factory):
         }
     }
 
+    rsp = cli.query_params("icaauth")
+    assert rsp["params"]["min_timeout_duration"] == "3600s", rsp
+
     # migrate to sdk v0.47
     custom_cronos.supervisorctl("stop", "all")
     sdk_version = "v0.47"