+++ x/cronos/events/decoders.go
@@ -6,6 +6,7 @@ import (
 	"errors"
 	"fmt"
 	"math/big"
+	"strconv"
 	"strings"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -22,7 +23,20 @@ type (
 	ValueDecoders map[string]ValueDecoder
 )
 
-const intBase = 10
+func (d ValueDecoders) GetDecoder(name string) (ValueDecoder, bool) {
+	decoder, ok := d[name]
+	if !ok {
+		decoder, ok = d[""]
+	}
+	return decoder, ok
+}
+
+const (
+	// intBase is the base `int`s are parsed in, 10.
+	intBase = 10
+	// int64Bits is the number of bits stored in a variabe of `int64` type.
+	int64Bits = 64
+)
 
 func AccAddressFromBech32(address string) (addr sdk.AccAddress, err error) {
 	if len(strings.TrimSpace(address)) == 0 {
@@ -111,6 +125,14 @@ func ConvertPacketData(attributeValue string, indexed bool) ([]any, error) {
 
 func ReturnStringAsIs(attributeValue string, _ bool) ([]any, error) {
 	return []any{attributeValue}, nil
+}
+
+func ConvertUint64(attributeValue string, _ bool) ([]any, error) {
+	res, err := strconv.ParseUint(attributeValue, intBase, int64Bits)
+	if err != nil {
+		return nil, err
+	}
+	return []any{res}, err
 }
 
 func convertAddress(addrString string) (*common.Address, error) {